Extract offer calculation into class of its own
